+++ packages/stream_chat_flutter/lib/src/localization/translations.dart
@@ -304,6 +304,8 @@ abstract class Translations {
 
   /// The label for "Reply to message"
   String get replyToMessageLabel;
+
+  String attachmentLimitExceedError(int limit);
 }
 
 /// Default implementation of Translation strings for the stream chat widgets
@@ -664,4 +666,8 @@ class DefaultTranslations implements Translations {
 
   @override
   String get replyToMessageLabel => 'Reply to Message';
+
+  @override
+  String attachmentLimitExceedError(int limit) =>
+      'Attachment limit exceeded, limit: $limit';
 }