+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -357,4 +357,8 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Reply to Message';
+
+  @override
+  String attachmentLimitExceedError(int limit) =>
+      'Attachment limit exceeded, limit: $limit';
 }