+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -362,4 +362,10 @@ class StreamChatLocalizationsEs extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Responder al Mensaje';
+
+  @override
+  String attachmentLimitExceedError(int limit) {
+    // TODO: implement attachmentLimitExceedError
+    throw UnimplementedError();
+  }
 }