+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -361,4 +361,10 @@ class StreamChatLocalizationsFr extends GlobalStreamChatLocalizations {
 
   @override
   String get replyToMessageLabel => 'Répondre au Message';
+
+  @override
+  String attachmentLimitExceedError(int limit) {
+    // TODO: implement attachmentLimitExceedError
+    throw UnimplementedError();
+  }
 }