+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_it.dart
@@ -358,4 +358,10 @@ Il file è troppo grande per essere caricato. Il limite è di $limitInMB MB.''';
 
   @override
   String get replyToMessageLabel => 'Rispondi al messaggio';
+
+  @override
+  String attachmentLimitExceedError(int limit) {
+    // TODO: implement attachmentLimitExceedError
+    throw UnimplementedError();
+  }
 }